feat: downIcon에도 애니메이션 적용 start page
