+++ src/pages/Start/StartAnimation/DownIconAnimation/StartDownIconAnimation.tsx
@@ -1,0 +1,14 @@
+import { ScrollAnimationUpContainer } from '../../../../components/ScrollAnimationContainer/ScrollAnimationContainerUp';
+import * as S from './StartDownIconAnimation.style';
+
+const StartDownIconAnimation = () => {
+  return (
+    <S.DownIconWrapper>
+      <ScrollAnimationUpContainer>
+        <S.DownIcon />
+      </ScrollAnimationUpContainer>
+    </S.DownIconWrapper>
+  );
+};
+
+export default StartDownIconAnimation;
